chore(refactor): refactor solana program for simplicity
