+++ programs/solana-twitter/src/constants.rs
@@ -1,0 +1,8 @@
+// 2. Add some useful constants for sizing propeties.
+pub const _CHAR_SIZE: usize = 4;
+pub const DISCRIMINATOR_LENGTH: usize = 8;
+pub const PUBLIC_KEY_LENGTH: usize = 32;
+pub const TIMESTAMP_LENGTH: usize = 8;
+pub const _STRING_LENGTH_PREFIX: usize = 4; // Stores the size of the string.
+pub const MAX_TOPIC_LENGTH: usize = 50; // 50 chars max.
+pub const MAX_CONTENT_LENGTH: usize = 280; // 280 chars max.
