+++ programs/solana-twitter/src/instructions/mod.rs
@@ -1,0 +1,2 @@
+pub mod delete_tweet;
+pub mod send_tweet;
